+++ HISTORY.md
@@ -124,7 +124,8 @@
 ### 2026-04-27
 * gui: made green-flag button change its color to red while any script has an error
 * blocks: guard piano keyboard dropdown menu against non-numerical values 
-* byob: rearranged some slot types in the long form input dialog so default "Any type" is the first one top left 
+* byob: rearranged some slot types in the long form input dialog so default "Any type" is the first one top left
+* store: tweaked loading local custom blocks to refresh custom dropdowns in variadic slots 
 
 ### 2026-04-23
 * blocks: trigger "When slot signals edited" for piano keyboard and direction dial dropdown menus